Sleep for a short time before connecting to Sigil2
Added a nanosleep syscall to give Sigil2 time to bring up
the interprocess communication fifos and shared memory region.

This had to be added because I removed the implicit guarantee
that the fifos and shared memory would be available before
the valgrind tool was forked.

If the fifos and shared memory do not exist by the timeout period,
then valgrind will error out.